+++ packages/pluggableWidgets/datagrid-web/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## [Unreleased]
 
 ### Added
+
+- We added "Tooltip" property for column, which allow you to control text that will be seen when hovering cell.
 - We added dark mode to Structure mode preview.
 
 ## [2.1.0] - 2021-12-3